% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_filter_list_to_sql}
\alias{from_filter_list_to_sql}
\title{Get formatted SQL filter from R named list}
\usage{
from_filter_list_to_sql(filter_list)
}
\arguments{
\item{filter_list}{Filter (named list).}
}
\description{
Get formatted SQL filter from R named list
}
\examples{

fl = list(year = '2018',industry = 'A01',serie_id = 'ghg_obs')
lsnstat:::from_filter_list_to_sql(fl)
}
