% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Get table-related endpoint}
\usage{
get_endpoint(dataset)
}
\arguments{
\item{dataset}{dataset requested.}
}
\description{
Get table-related endpoint
}
\examples{

lsnstat:::get_endpoint("figaro_main_aggregates")
}
