% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnstat.R
\name{lsnrstat}
\alias{lsnrstat}
\title{R companion of La Societe Nouvelle macro_data API services}
\usage{
lsnrstat(
  table = c("footprint", "structural"),
  activity = "",
  year = "",
  indic = c("", "ECO", "ART", "GHG", "SOC", "GEQ", "NRG", "MAT", "HAZ", "KNW", "WAT",
    "WAS", "IDR"),
  area = "",
  type = c("", "tgt", "trd", "cpeb", "ere", "pat_nf", "tei", "tess"),
  classification = c("A38", "A88"),
  flag = "",
  path = c("", "linear", "geometric"),
  currency = "",
  product = "",
  unit
)
}
\arguments{
\item{table}{Kind of data requested : eco-environmental footprints or structural forecasted economic data (required)}

\item{activity}{NACE Rev.2 economic activity requested : branch or division code (optional)}

\item{year}{Year or years requested (optional)}

\item{indic}{For footprint requests, eco-environnemental footprint dimension (optional). A list is available in the lsnr package \code{\link[lsnr:get_indicator_list]{lsnr::get_indicator_list()}}}

\item{area}{Geographic filter (optional)}

\item{type}{For footprint requests, type of data : observed, target, trend (required for footprints request)}

\item{classification}{Data NACE Rev.2 granularity : branch by 38 or division by 88 (required for footprints request)}

\item{flag}{Flag filter (optional). It informs on the status of each data observation.}

\item{path}{Linear or geometric path for target aims to be fulfilled (optional)}

\item{currency}{Currency filter (optional). Allows to narrow the request to current or constant prices.}

\item{product}{Only for structural economic data, filter for product (optional)}

\item{unit}{Unit filter (optional)}
}
\value{
A \code{\link[=data.frame]{data.frame()}}.
}
\description{
R companion of La Societe Nouvelle macro_data API services
}
\examples{

#Get CPEB forecasted data for division 10 between 2022 and 2030.

lsnrstat(table = "structural",year = 2022:2030,type = "cpeb",activity = "10")

#Get TEI forecasted data for branch JA use of product OZ between 2022 and 2030.

lsnrstat(table = "structural",year = 2022:2030,type = "tei",activity = "JA",product = "OZ")

#Get branch EZ 2018 footprints for NRG, GHG and ECO indicators.

lsnrstat(table = "footprint",year = 2018,activity = "EZ",indic = c("NRG","GHG","ECO"))

}
