% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnstat-package.R
\docType{package}
\name{lsnstat-package}
\alias{lsnstat}
\alias{lsnstat-package}
\title{lsnstat: La Societe Nouvelle API access}
\description{
The lsnstat package was created to facilitate access to the macro_data service of the La Societe Nouvelle API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/La-Societe-Nouvelle}
  \item \url{https://lasocietenouvelle.org/}
  \item \url{https://api.lasocietenouvelle.org/,https://lasocietenouvelle.org/portail}
  \item \url{https://www.linkedin.com/company/la-societe-nouvelle/.}
}

}
\author{
\strong{Maintainer}: Joris BLAIN \email{joris.blain@lasocietenouvelle.org}

}
\keyword{internal}
