% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnstat_endpoint.R
\name{lsnstat_endpoint}
\alias{lsnstat_endpoint}
\title{Link La Societe Nouvelle API services and R for flexible endpoints : return raw table}
\usage{
lsnstat_endpoint(endpoint, dataset = NULL, filters = NULL)
}
\arguments{
\item{endpoint}{endpoint requested : list available at \url{https://api.lasocietenouvelle.org/} (required)}

\item{dataset}{dataset requested (optional)}

\item{filters}{filters to apply (optional)}
}
\description{
Link La Societe Nouvelle API services and R for flexible endpoints : return raw table
}
\examples{

#macrodata macro_fpt request

lsnstat_endpoint(endpoint = "macrodata", dataset = "macro_fpt", filters = "industry=C20&indic=GHG+NRG&year=2020")

}
