% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnstat_getdata.R
\name{lsnstat_getdata}
\alias{lsnstat_getdata}
\title{R generic companion of 'La Societe Nouvelle' for all API services}
\usage{
lsnstat_getdata(dataset, filters, service = "macrodata")
}
\arguments{
\item{dataset}{dataset requested : list available at \url{https://docs.lasocietenouvelle.org/series-donnees} (required)}

\item{filters}{filters to apply : list of dataset 'params' available through service \link{lsnstat_metadata} (optional)}

\item{service}{service requested : list available at \url{https://api.lasocietenouvelle.org/} (required)}
}
\value{
A \code{\link[=data.frame]{data.frame()}}.
}
\description{
R generic companion of 'La Societe Nouvelle' for all API services
}
\examples{
# See [lsnstat_macrodata] and [lsnstat_statsdata] for precise examples.

}
