% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnstat_macrodata.R
\name{lsnstat_macrodata}
\alias{lsnstat_macrodata}
\title{R companion of 'La Societe Nouvelle' macrodata API services}
\usage{
lsnstat_macrodata(dataset, filters = NULL)
}
\arguments{
\item{dataset}{dataset requested : list available at \url{https://docs.lasocietenouvelle.org/series-donnees} (required)}

\item{filters}{filters to apply : list of dataset 'params' available through service \link{lsnstat_metadata} (optional)}
}
\value{
A \code{\link[=data.frame]{data.frame()}}.
}
\description{
R companion of 'La Societe Nouvelle' macrodata API services
}
\examples{

# GET industry 'A01' 2018 production footprint for 'NRG' and 'GHG' indicators.

lsnstat_macrodata(
  dataset = "macro_fpt",
  filters="industry=A01&year=2018&aggregate=PRD&indic=NRG+GHG"
  )

}
