% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnstat_statdata.R
\name{lsnstat_statdata}
\alias{lsnstat_statdata}
\title{R companion of 'La Societe Nouvelle' statdata API services}
\usage{
lsnstat_statdata(dataset, filters)
}
\arguments{
\item{dataset}{dataset requested (required)}

\item{filters}{filters to apply (optional)}
}
\value{
A \code{\link[=data.frame]{data.frame()}}.
}
\description{
R companion of 'La Societe Nouvelle' statdata API services
}
\examples{

# GET industry 'K66' 2018 direct impacts over production value for 'NRG' and 'GHG' indicators.

lsnstat_statdata(
  dataset = "direct_impacts_production",
  filters="id=FR_K66&year=2018&&indic=NRG+GHG"
  )

}
