% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnstat_statsdata.R
\name{lsnstat_statsdata}
\alias{lsnstat_statsdata}
\title{R companion of 'La Societe Nouvelle' stats_data API service}
\usage{
lsnstat_statsdata(dataset, filters)
}
\arguments{
\item{dataset}{dataset requested : list available at ... (required)}

\item{filters}{filters to apply : list of dataset 'params' available at ... (optional)}
}
\value{
A \code{\link[=data.frame]{data.frame()}}.
}
\description{
R companion of 'La Societe Nouvelle' stats_data API service
}
\examples{

# GET the value added of sector C19 of France in 2018.

lsnstat_statsdata(
  dataset = "figaro_main_aggregates",
  filters="year=2018&aggregate=VA&country=FR&industry=C19"
  )

}
