% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnstat-package.R
\docType{package}
\name{lsnstat-package}
\alias{lsnstat}
\alias{lsnstat-package}
\title{lsnstat: 'La Societe Nouvelle' API Access}
\description{
Tools facilitating access to the 'macro_data' service of the 'La Societe Nouvelle' API. It ensures an easy and fully-disclosed access to all macro-level data used in the 'La Societe Nouvelle' systems and the related metadata. Related API can be accessed from \url{https://api.lasocietenouvelle.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/La-Societe-Nouvelle/lsnstat/}
  \item Report bugs at \url{https://github.com/La-Societe-Nouvelle/lsnstat/issues/}
}

}
\author{
\strong{Maintainer}: Joris Blain \email{joris.blain@lasocietenouvelle.org}

}
\keyword{internal}
